Move mock VGL functions to util.
This will make in convenient for other projects to test geojs in non-webgl browsers.
