+++ src/util/index.js
@@ -5,6 +5,7 @@ var $ = require('jquery');
  */
 var util = require('./init');
 $.extend(util, require('./throttle'));
+$.extend(util, require('./mockVGL'));
 util.DistanceGrid = require('./distanceGrid.js');
 util.ClusterGroup = require('./clustering.js');
