+++ tests/cases/pointFeature.js
@@ -1,13 +1,13 @@
 // Test geo.pointFeature, geo.d3.pointFeature, and geo.gl.pointFeature
 
-var geo = require('../test-utils').geo;
 var $ = require('jquery');
 var mockAnimationFrame = require('../test-utils').mockAnimationFrame;
 var stepAnimationFrame = require('../test-utils').stepAnimationFrame;
 var unmockAnimationFrame = require('../test-utils').unmockAnimationFrame;
+var geo = require('../test-utils').geo;
+var mockVGLRenderer = geo.util.mockVGLRenderer;
+var restoreVGLRenderer = geo.util.restoreVGLRenderer;
 var vgl = require('vgl');
-var mockVGLRenderer = require('../test-utils').mockVGLRenderer;
-var restoreVGLRenderer = require('../test-utils').restoreVGLRenderer;
 var waitForIt = require('../test-utils').waitForIt;
 
 describe('geo.pointFeature', function () {
